backend/user_service/requirements.txt
fastapi
SQLAlchemy
pydantic
pydantic[email]
passlib[bcrypt]
python-jose
psycopg2-binary
uvicorn
PyJWT 